[future][core] extended jobStateEvent event listening
