+++ docs/en/concept/event-listener.md
@@ -141,7 +141,7 @@ Introduce the necessary dependencies in the project's `pom.xml`:
 </dependency>
 <dependency>
     <groupId>org.apache.seatunnel</groupId>
-    <artifactId>seatunnel-engine-server</artifactId>
+    <artifactId>seatunnel-engine-common</artifactId>
     <version>${seatunnel.version}</version>
     <scope>provided</scope>
 </dependency>
@@ -156,8 +156,8 @@ Create a custom class that implements the `org.apache.seatunnel.api.event.EventH
 import lombok.extern.slf4j.Slf4j;
 import org.apache.seatunnel.api.event.Event;
 import org.apache.seatunnel.api.event.EventHandler;
-import org.apache.seatunnel.engine.core.job.JobStatus;
-import org.apache.seatunnel.engine.server.event.JobStateEvent;
+import org.apache.seatunnel.engine.common.job.JobStatus;
+import org.apache.seatunnel.engine.common.job.JobStateEvent;
 
 /**
  * Custom Job status event handler, example includes log recording and exception alerting logic
