+++ seatunnel-engine/seatunnel-engine-common/src/main/java/org/apache/seatunnel/engine/common/job/JobStatusData.java
@@ -15,23 +15,22 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.engine.core.job;
+package org.apache.seatunnel.engine.common.job;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NonNull;
+import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
 
-@Data
 @AllArgsConstructor
-public class JobResult implements Serializable {
-
-    @NonNull private JobStatus status;
-
-    private String error;
-
-    public JobResult(@NonNull JobStatus status) {
-        this.status = status;
-    }
+@Data
+@NoArgsConstructor
+public final class JobStatusData implements Serializable {
+    private Long jobId;
+    private String jobName;
+    private JobStatus jobStatus;
+    private long submitTime;
+    private Long startTime;
+    private Long finishTime;
 }
