+++ seatunnel-engine/seatunnel-engine-server/src/test/java/org/apache/seatunnel/engine/server/event/JobStateEventTest.java
@@ -20,7 +20,8 @@ package org.apache.seatunnel.engine.server.event;
 import org.apache.seatunnel.api.event.EventHandler;
 import org.apache.seatunnel.api.event.EventType;
 import org.apache.seatunnel.common.utils.ReflectionUtils;
-import org.apache.seatunnel.engine.core.job.JobStatus;
+import org.apache.seatunnel.engine.common.job.JobStateEvent;
+import org.apache.seatunnel.engine.common.job.JobStatus;
 import org.apache.seatunnel.engine.server.AbstractSeaTunnelServerTest;
 
 import org.junit.jupiter.api.Assertions;
@@ -29,6 +30,7 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicReference;
 
 import static org.apache.seatunnel.engine.server.checkpoint.CheckpointErrorRestoreEndTest.STREAM_CONF_WITH_ERROR_PATH;
 import static org.awaitility.Awaitility.await;
@@ -42,18 +44,21 @@ public class JobStateEventTest extends AbstractSeaTunnelServerTest {
                 (JobEventProcessor) server.getCoordinatorService().getEventProcessor();
 
         AtomicInteger accessCounter = new AtomicInteger(0);
+        AtomicReference<JobStateEvent> jobStateEventReference = new AtomicReference<>();
         EventHandler eventHandler =
                 event -> {
                     if (event.getEventType() != EventType.JOB_STATUS) {
                         return;
                     }
-                    JobStatus status = ((JobStateEvent) event).getJobStatus();
+                    JobStateEvent jobStateEvent = (JobStateEvent) event;
+                    JobStatus status = jobStateEvent.getJobStatus();
                     switch (status) {
                         case FAILED:
                         case CANCELED:
                         case SAVEPOINT_DONE:
                         case FINISHED:
                             accessCounter.incrementAndGet();
+                            jobStateEventReference.lazySet(jobStateEvent);
                             break;
                         default:
                             break;
@@ -64,8 +69,9 @@ public class JobStateEventTest extends AbstractSeaTunnelServerTest {
                 (List<EventHandler>) ReflectionUtils.getField(eventProcessor, "handlers").get();
         handlers.add(eventHandler);
         long jobId_finished = System.currentTimeMillis();
+        long currentTimeMillis = System.currentTimeMillis();
         startJob(jobId_finished, "fake_to_console.conf", false);
-        await().atMost(120000, TimeUnit.MILLISECONDS)
+        await().atMost(60, TimeUnit.SECONDS)
                 .untilAsserted(
                         () ->
                                 Assertions.assertEquals(
@@ -74,10 +80,15 @@ public class JobStateEventTest extends AbstractSeaTunnelServerTest {
                                                 .getJobStatus(jobId_finished)));
         // check whether the event handler is executed
         Assertions.assertEquals(1, accessCounter.get());
+        JobStateEvent jobStateEventFinished = jobStateEventReference.get();
+        Assertions.assertEquals(String.valueOf(jobId_finished), jobStateEventFinished.getJobId());
+        Assertions.assertEquals(JobStatus.FINISHED, jobStateEventFinished.getJobStatus());
+        Assertions.assertTrue(jobStateEventFinished.getCreatedTime() > currentTimeMillis);
+        Assertions.assertEquals(String.valueOf(jobId_finished), jobStateEventFinished.getJobName());
 
         long jobId_failed = System.currentTimeMillis();
         startJob(jobId_failed, STREAM_CONF_WITH_ERROR_PATH, false);
-        await().atMost(120, TimeUnit.SECONDS)
+        await().atMost(60, TimeUnit.SECONDS)
                 .untilAsserted(
                         () ->
                                 Assertions.assertEquals(
@@ -85,5 +96,10 @@ public class JobStateEventTest extends AbstractSeaTunnelServerTest {
                                         server.getCoordinatorService().getJobStatus(jobId_failed)));
 
         Assertions.assertEquals(2, accessCounter.get());
+        JobStateEvent jobStateEventFailed = jobStateEventReference.get();
+        Assertions.assertEquals(String.valueOf(jobId_failed), jobStateEventFailed.getJobId());
+        Assertions.assertEquals(JobStatus.FAILED, jobStateEventFailed.getJobStatus());
+        Assertions.assertTrue(jobStateEventFailed.getCreatedTime() > currentTimeMillis);
+        Assertions.assertEquals(String.valueOf(jobId_failed), jobStateEventFailed.getJobName());
     }
 }
